changepassword component, Toast in commonFuncs
