+++ frontend/src/commonFuncs.js
@@ -1,4 +1,5 @@
 import { SERVER_URL } from "./config";
+import Swal from 'sweetalert2';
 
 export const getDomains = async() => {
   var requestOptions = {
@@ -41,3 +42,11 @@ export const getOptionsForYear = () => {
   }
   return data;
 }
+
+export const Toast = Swal.mixin({
+  toast: true,
+  position: 'top-end',
+  showConfirmButton: false,
+  timer: 3000,
+  timerProgressBar: true,
+})
