+++ frontend/src/components/StudentEdit.js
@@ -3,6 +3,7 @@ import { Grid, TextField, FormControl, InputLabel, MenuItem, Select, makeStyles,
 import { SERVER_URL } from '../config';
 import { githubValid } from './Upload';
 import Swal from 'sweetalert2';
+import ChangePassword from './ChangePassword';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -19,15 +20,7 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-const Toast = Swal.mixin({
-  toast: true,
-  position: 'top-end',
-  showConfirmButton: false,
-  timer: 4000,
-  timerProgressBar: true,
-})
-
-function StudentEdit() {
+const StudentEdit = () => {
   const [githubId, setgithubId] = useState(
     localStorage.getItem("githubId")?
     localStorage.getItem("githubId"):
@@ -37,45 +30,6 @@ function StudentEdit() {
   const [year, setYear] = useState(localStorage.getItem("year"));
   const [error, seterror] = useState(null);
   const classes = useStyles();
-  const [currentPass, setCurrentPass] = useState("");
-  const [newPass, setNewPass] = useState("");
-
-  const changePass = (props) => {
-    console.log(currentPass, newPass)
-    const word = 'Token ';
-    const token = word.concat(`${localStorage.getItem('Token')}`);
-    var myHeaders = new Headers();
-    myHeaders.append('Authorization', `${token}`);
-    var formdata = new FormData();
-    formdata.append("current_password", currentPass);
-    formdata.append("new_password", newPass);
-    var requestOptions = {
-      method: 'POST',
-      headers: myHeaders,
-      body: formdata,
-      redirect: 'follow'
-    }
-    fetch(`${SERVER_URL}/change_password`, requestOptions)
-      .then(result => {
-        console.log(result)
-        if (result.status === 204) {
-          Toast.fire({
-            icon: 'success',
-            title: 'Password changed successfully'
-          })
-          setNewPass("")
-          setCurrentPass("")
-        }
-        if(result.status === 400){
-          Toast.fire({
-            icon: 'error',
-            title: 'Invalid Credentials'
-          })
-          setNewPass("")
-          setCurrentPass("")
-        }
-      })  
-  }
 
   const checkError = (url) => {
     if(githubValid.test(url)) {
@@ -185,45 +139,7 @@ function StudentEdit() {
           Change
         </Button>
       </div>
-      <Grid container spacing={4} className={classes.title} >
-        <Grid md={12} xs={12}>
-          <h6>Change Password</h6>
-        </Grid>
-      </Grid>
-      <Grid container spacing={ 4 }>
-        <Grid item md={ 12 } xs={ 12 }>
-          <TextField
-            id="currentPass"
-            fullWidth
-            type="password"
-            value={ currentPass }
-            label="Current Password"
-            onChange={ e => setCurrentPass(e.target.value) }
-          />
-        </Grid>
-      </Grid>
-      <Grid container spacing={ 4 }>
-        <Grid item md={ 12 } xs={ 12 }>
-          <TextField
-            id="newPass"
-            fullWidth
-            type="password"
-            value={ newPass }
-            label="New Password"
-            onChange={ e => setNewPass(e.target.value) }
-          />
-        </Grid>
-      </Grid>
-      <div>
-        <Button
-          variant="contained"
-          color="primary"
-          onClick={ changePass }
-          className={ classes.changeButton }
-        >
-          Change
-        </Button>
-      </div>
+      <ChangePassword classes={classes}/>
     </div>
   )
 }
